requirements.txt
PyQt6==6.6.1
Pillow==10.2.0
pynput==1.7.6
pyautogui==0.9.54
numpy==1.26.4
reportlab==4.1.0
python-docx==1.1.0 